commentaire avec code qui marche pas
storage aad.xaml.cs et storge controller

+++ Models/Storage.model.cs
@@ -6,14 +6,14 @@ namespace EquiLog.Models
     {
         public int Saddle { get; set; }
         public int Briddle { get; set; }
-        public double Granule { get; set; }
+        public int Granule { get; set; }
         public int Blanket { get; set;}
-        public double Hay { get; set; }
+        public int Hay { get; set; }
         public int SaddlePad { get; set; }
         public int StirrupStraps { get; set; }
         public int Stirrup { get; set; }
         public string OrderDemand { get;set; }
-        public Storage_model(double hay,int saddle, int briddle, double granule, int blanket, int saddlepad, int stirrupstraps, int stirrup,string orderdemand)
+        public Storage_model(int hay,int saddle, int briddle, int granule, int blanket, int saddlepad, int stirrupstraps, int stirrup,string orderdemand)
         {
             Hay = hay;
             Saddle = saddle;
